requirements.txt
absl-py==0.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
albumentations==1.0.0
altair==4.2.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
av==12.3.0
backports.zoneinfo==0.2.1
blinker==1.8.2
Bottleneck @ file:///croot/bottleneck_1707864210935/work
box==0.1.5
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6
Columnar==1.3.1
commonmark==0.9.1
contourpy==1.1.1
cryptography==44.0.2
cycler==0.12.1
dlib==19.24.6
dnspython==2.6.1
entrypoints==0.4
executing==0.8.2
flatbuffers==1.12
fonttools==4.56.0
frozenlist==1.5.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.10
ifaddr==0.2.0
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
loguru==0.7.3
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.4.3
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
multidict==6.1.0
networkx==3.1
numexpr @ file:///croot/numexpr_1683221822650/work
numpy==1.21.0
oauthlib==3.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==24.2
pandas==1.4.4
pillow==10.4.0
pkgutil_resolve_name==1.3.10
propcache==0.2.0
protobuf==3.20.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
pylibsrtp==0.10.0
Pympler==1.1
pyOpenSSL==24.3.0
pyparsing==3.1.4
pysqlite3-binary==0.5.4
python-box==6.1.0
python-dateutil==2.8.2
pytz @ file:///croot/pytz_1713974312559/work
PyWavelets==1.4.1
PyYAML==6.0.2
qudida==0.0.4
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retinaface==1.1.1
retinaface-pytorch==0.0.8
rich==12.0.1
rpds-py==0.20.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==1.3.2
scipy==1.10.1
semver==3.0.4
shapely==2.0.7
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
streamlit==1.10.0
streamlit-webrtc==0.42.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timeago==1.0.14
toml==0.10.2
toolz==1.0.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.4.2
tqdm @ file:///croot/tqdm_1724853939799/work
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==4.0.2
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.12.1
yarl==1.15.2
zipp==3.20.2

